fix: change component layout and removed outside click hook
