+++ src/Router.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Layout from './components/Layout';
+import Layout from './components/layout/Layout';
 import Error404 from './routes/Error404';
 import Home from './routes/Home';
 import Resume from './routes/Resume';
