+++ src/components/layout/Header.tsx
@@ -1,11 +1,30 @@
-import React, { useState } from 'react';
+import React, { useEffect, useRef, useState } from 'react';
 import { Link } from 'react-router-dom';
-import { useClickOutsideRef } from '../hooks';
-
 
 const Header: React.FC = () => {
     const [ isNavbarOpen, setIsNavbarOpen ] = useState<boolean>(false);
-    const clickOutsideNavbarRef = useClickOutsideRef<HTMLButtonElement>(() => setIsNavbarOpen(false));
+    const navbarRef = useRef<HTMLButtonElement>(null);
+
+    useEffect(() => {
+        if (navbarRef.current === null) {
+            return;
+        }
+
+        const handleClick = (event: Event) => {
+            if (!navbarRef.current?.contains(event.target as Node)) {
+                setIsNavbarOpen(false);
+            }
+        };
+
+        document.addEventListener('click', handleClick);
+        return () => document.removeEventListener('click', handleClick);
+    }, []);
+
+    const handleNavbarLinkClick = (event: React.MouseEvent) => {
+        event.stopPropagation();
+        setIsNavbarOpen(false);
+    };
+
 
     return (
         <div className='flex-none py-10'>
@@ -18,7 +37,7 @@ const Header: React.FC = () => {
                 <Link className='h-fit hover:underline hover:underline-offset-2' to='/resume'>/resume</Link>
             </div>
             <div className="flex flex-row justify-between md:hidden">
-                <button ref={ clickOutsideNavbarRef } onClick={ () => setIsNavbarOpen(!isNavbarOpen) }>
+                <button ref={ navbarRef } onClick={ () => setIsNavbarOpen(!isNavbarOpen) }>
                     { isNavbarOpen
                         ? (
                             <img
@@ -42,13 +61,13 @@ const Header: React.FC = () => {
             </div>
             <div
                 className={ isNavbarOpen
-                    ? 'absolute my-2 flex w-2/5 flex-col space-y-0 border-2 border-dotted bg-white align-middle'
+                    ? 'absolute z-50 my-2 flex w-2/5 flex-col space-y-0 border-2 border-dotted bg-white align-middle'
                     : 'hidden' }
             >
-                <Link onClick={ () => setIsNavbarOpen(false) } to='/' className='border-b-2 border-dotted p-2'>$HOME</Link>
-                <Link onClick={ () => setIsNavbarOpen(false) } to='/blog' className='border-b-2 border-dotted p-2'>/blog</Link>
-                <Link onClick={ () => setIsNavbarOpen(false) } to='/projects' className='border-b-2 border-dotted p-2'>/projects</Link>
-                <Link onClick={ () => setIsNavbarOpen(false) } to='/resume' className='p-2'>/resume</Link>
+                <Link onClick={ handleNavbarLinkClick } to='/' className='border-b-2 border-dotted p-2'>$HOME</Link>
+                <Link onClick={ handleNavbarLinkClick } to='/blog' className='border-b-2 border-dotted p-2'>/blog</Link>
+                <Link onClick={ handleNavbarLinkClick } to='/projects' className='border-b-2 border-dotted p-2'>/projects</Link>
+                <Link onClick={ handleNavbarLinkClick } to='/resume' className='p-2'>/resume</Link>
             </div>
         </div>
     );
